refactored tests for constructive heuristics
